Git - use random name for copilot worktree branch names
